+++ src/App.js
@@ -1,6 +1,13 @@
 import React from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+
+/* import components */
+import Header from './components/Header';
 import Calculator from './components/calculator/calculator';
-import Result from './components/calculator/result';
+import Home from './pages/Home';
+import Quote from './pages/Quote';
+
+import './App.css';
 
 // eslint-disable-next-line react/prefer-stateless-function
 class App extends React.Component {
@@ -20,9 +27,20 @@ class App extends React.Component {
     const { result } = this.state;
 
     return (
-      <div>
-        <Result result={result} />
-        <Calculator updateResult={this.updateResult} />
+      <div className="container">
+        <Router>
+          <Header />
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route
+              path="/calculator"
+              element={
+                <Calculator updateResult={this.updateResult} result={result} />
+              }
+            />
+            <Route path="/quote" element={<Quote />} />
+          </Routes>
+        </Router>
       </div>
     );
   }
