+++ src/pages/Home.js
@@ -1,0 +1,27 @@
+import animation from '../assets/images/bg.png';
+
+const Home = () => (
+  <div>
+    <p>
+      What is a calculator? A calculator is a device that performs arithmetic
+      operations on numbers. Basic calculators can do only addition,
+      subtraction, multiplication and division mathematical calculations.
+      However, more sophisticated calculators can handle exponential operations,
+      square roots, logarithms, trigonometric functions and hyperbolic
+      functions. Internally, some calculators perform all these functions by
+      repeated addition processes.
+    </p>
+    <img src={animation} alt="background main page" className="main-page-img" />
+    <p>
+      The evolution of the calculator Most calculators these days require
+      electricity to operate or are battery-powered calculators. Calculators
+      work by performing programmed functions based on numerical inputs. Before
+      the electronic calculator (circa 1970), a more primitive calculator, the
+      slide rule, was commonly used. It consisted of a slat of wood called the
+      slide that could be moved in and out of a reinforced pair of slats. Both
+      the slide and the outer pair of slats had calibrated numerical scales.
+    </p>
+  </div>
+);
+
+export default Home;
