+++ src/pages/Quote.js
@@ -1,0 +1,5 @@
+const Quote = () => (
+  <h3>Comming soon .....</h3>
+);
+
+export default Quote;
